# Evaluate `UTxOIndex` value to WHNF outside of `transactionFee` loop.