@piotr-iohk 2023-04-26
Testing the same wallet on `preprod`:

```
time curl -X POST http://localhost:8090/v2/wallets/1f82e83772b7579fc0854bd13db6a9cce21ccd95/payment-fees -d '{"payments":[{"address":"addr_test1qrtez7vn0d8xp495ggypmu2kyt7tt6qyva2spm0f5a3ewn0v474mcs4q8e9g55yknx3729kyg5dl69x5596ee9tvnynq7ffety","amount":{"quantity":100000000,"unit":"lovelace"}}]}' -H "Content-Type: application/json"
```



v2023-04-14 (git revision: 3f9055373c272a4d7b1f9f6bacc1df3b6dc906af) (release)

```
0,116s
0,067s
0,251s
0,063s
0,111s
```

v2023-04-14 (git revision: cec0640ceac3ddcdfd92c4efc02f31dcf3287b7d) (current master)

```
0,524s
0,299s
0,325s
0,375s
0,287s
```


v2023-04-14 (git revision: 0a0b93465d2be3da44eefd02a285250b66ef3cc0) (this branch)

```
0,265s
0,257s
0,172s
0,186s
0,210s
```