@HeinrichApfelmus 2023-05-02
> Therefore, evaluating a value of type UTxOIndex to WHNF should be enough to ensure that all constituent fields are fully evaluated, assuming of course that there are no unevaluated thunks within the internal maps.

Alas, this is not the actual reason why performance improves 😅  — the actual reason is that `evaluate` prevents the GHC state hack from tacking place, presumably because `evaluate` specifies an order in which the think `utxo` is allocated.

See also the following: [StackOverflow answer — What is the state hack, and why does it destroy my programs performance](https://stackoverflow.com/a/30603291). The `nestedArray` from that example is very much like `utxo` here.

This implies that the following changes would also have improved performance:

1. Using a monad other than `IO` for the `MonadRandom` in `calculateFeePercentiles`.
2. Using a WHNF that is very far from an NF, e.g. `evaluate $ Just . UTxOIndex.fromMap . …`.